pyproject.toml
[project]
name = "clipmind-backend"
version = "0.1.0"
description = "AI-powered video clipping and content repurposing platform."
authors = [{name = "ClipMind Team"}]
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "celery>=5.3.6",
    "redis>=5.0.4",
    "pydantic>=2.7.1",
    "python-multipart>=0.0.9",
    "openai>=1.30.1",
    "sqlalchemy>=2.0.30",
    "alembic>=1.13.1",
    "psycopg2-binary>=2.9.9",
    "ffmpeg-python>=0.2.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
    "tenacity>=8.3.0",
    "requests>=2.31.0",
    "yt-dlp>=2026.03.17",
    "librosa>=0.11.0",
    "watchfiles>=0.21.0",
    "google-api-python-client>=2.115.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
    "sentry-sdk[fastapi]>=2.3.1",
    "supabase>=2.4.6",
    "stripe>=8.8.0",
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.6",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
